Fix image bug when moving players over empty spots
There was an issue with moving a player to an empty spot,
where the image would change into the default. This was
slightly odd and a definitive explanation was never found.
It seems to have been related to that we changed the image
source when the enabled field changed, as when that feature
was removed, it worked as expected. Hm. Anyhow, changing it
to a multi-value converter resolved the issue. An obvious
improvment we can make is to try and make use of some caching
to not recreate the images whenever a player is moved, but let's
do that in a later patch.