Refactor layout logic
Before, we used the visibility trick to conditionally hide different
layouts. This caused us to actually generate all components on
startup, which wasn't very nice. The more correct WPF-way to do
this is to use the DataTemplate feature, which is sort of made
for this type of thing.

Now, we only create the current layout. Moreover, we adjust the
layouts so that they all look nice (at least in 1920x1080).